+++ ricaun.RevitTest.TestAdapter/TestExecutor.cs
@@ -148,6 +148,12 @@ namespace ricaun.RevitTest.TestAdapter
                     }
                 };
 
+                Action<string> outputError = (item) =>
+                {
+                    if (string.IsNullOrEmpty(item)) return;
+                    AdapterLogger.Logger.Error($"OutputConsole: ERROR: {item}");
+                };
+
                 AdapterLogger.Logger.Info($"RunRevitTest: {source} [Version: {AdapterSettings.Settings.NUnit.Version}] [TestFilter: {filters.Length}]");
 
                 await revit.RunTestAction(source,
@@ -155,7 +161,7 @@ namespace ricaun.RevitTest.TestAdapter
                     AdapterSettings.Settings.NUnit.Language,
                     AdapterSettings.Settings.NUnit.Open,
                     AdapterSettings.Settings.NUnit.Close,
-                    outputConsole, filters);
+                    outputConsole, outputError, filters);
             }
 
             AdapterLogger.Logger.Info("---------");